requirements.txt
playwright==1.51.0
pypdf==5.4.0
python-dotenv==1.0.1
pymysql==1.1.1
ruff==0.11.0
